refactor(actor): factor mailbox queue logic
